backend/requirements.txt
asgiref==3.5.2
Django==4.1.2
sqlparse==0.4.3
python-dotenv==0.21.0
psycopg2-binary==2.9.3
Pillow==9.2.0
djangorestframework==3.14.0
pytz==2022.4
certifi==2022.9.24
charset-normalizer==2.1.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.4
itypes==1.2.0
pycparser==2.21
pyjwt==2.5.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.12
gunicorn